+++ tests/02_eradiate/01_unit/scenes/atmosphere/test_molecular_atmosphere.py
@@ -59,8 +59,7 @@ def test_molecular_atmosphere_afgl_1986(mode_ckd, bin):
     dictionary."""
     bin = eradiate.scenes.measure._core.CKDMeasureSpectralConfig(bins=bin).bins[0]
     bindex = eradiate.ckd.Bindex(bin=bin, index=3)
-    spectral_ctx = eradiate.contexts.CKDSpectralContext(bindex=bindex, bin_set="10nm")
-    ctx = KernelDictContext(spectral_ctx=spectral_ctx)
+    ctx = KernelDictContext(spectral_ctx={"bindex": bindex, "bin_set": "10nm"})
     atmosphere = MolecularAtmosphere.afgl_1986(geometry="plane_parallel")
     assert atmosphere.kernel_dict(ctx).load()