+++ CHANGELOG.md
@@ -45,6 +45,7 @@ when necessary—we also advise to not ignore `DeprecationWarning`s.
 * Add support for all missing AFGL 1986 reference atmospheres in CKD mode ({ghpr}`185`).
 * Fix incorrect phase function blending in multi-component atmospheres ({ghpr}`197`).
 * Fix incorrect volume data transform for spherical heterogeneous atmospheres ({ghpr}`199`).
+* Add default value for `CKDSpectralContext.bin_set` ({ghpr}`205`).
 
 % ### Documentation